fix(components): stop leaking LiquidJS Node types in public declarations
Remove createLiquidEngine from the public barrel (it's an internal
<zitadel-login> rendering detail) and move TEMPLATE_NAMES into its own
liquidjs-free module. Together this keeps liquid.d.ts and its
`import { Liquid } from "liquidjs"` out of the published .d.mts, so
consumers with skipLibCheck:false no longer hit TS2688/TS2503 from
LiquidJS' Node ambient types.
