+++ packages/components/src/orchestrator/index.ts
@@ -16,7 +16,15 @@ export {
 } from "./branding-to-tokens.js";
 export { ThemeController, type ResolvedTheme } from "./theme-controller.js";
 export { applyFontUrl } from "./font-loader.js";
-export { createLiquidEngine, TEMPLATE_NAMES } from "./liquid.js";
+// `createLiquidEngine` is intentionally NOT re-exported: it returns LiquidJS'
+// `Liquid` type, whose declarations reference Node ambient types (`NodeJS`),
+// which would force every browser consumer of this package to install
+// `@types/node` under `skipLibCheck: false`. It's an internal rendering detail
+// of `<zitadel-login>` — import it directly from `./liquid.js` within the
+// package. `TEMPLATE_NAMES` is sourced from its own liquidjs-free module for
+// the same reason (re-exporting it from `liquid.js` would pull that `Liquid`
+// import back into the public declaration bundle).
+export { TEMPLATE_NAMES } from "./template-names.js";
 export { en, de, builtinLocales, type Locale } from "./locales/index.js";
 export {
   patchMandatoryGates,
