+++ packages/components/src/orchestrator/zitadel-login.ts
@@ -22,7 +22,8 @@ import type { Branding } from "./branding.js";
 import { applyBaseTokens, applyBrandingTokens } from "./branding-to-tokens.js";
 import { validateBranding } from "./branding-validator.js";
 import { applyFontUrl } from "./font-loader.js";
-import { createLiquidEngine, TEMPLATE_NAMES } from "./liquid.js";
+import { createLiquidEngine } from "./liquid.js";
+import { TEMPLATE_NAMES } from "./template-names.js";
 import { en, builtinLocales, type Locale } from "./locales/index.js";
 import { patchMandatoryGates } from "./mandatory-gates.js";
 import { zitadelAttributionPillInnerHtml } from "@zitadel/shared-component-styles/attribution-markup";
